+++ R/utils.R
@@ -1,6 +1,17 @@
-# set up zoltar connection
-setup_zoltar_connection <- function() {
-  zoltar_connection <- zoltr::new_connection()
+#' Set up Zoltar connection
+#' 
+#' @param staging boolean to change Zoltar server for staging
+#' 
+#' @return a Zoltar connection
+#' @export
+setup_zoltar_connection <- function(staging = FALSE) {
+  if(staging){
+    host = "https://rl-zoltar-staging.herokuapp.com"
+  } else {
+    host = "https://zoltardata.com"
+  }
+  
+  zoltar_connection <- zoltr::new_connection(host)
   if(Sys.getenv("Z_USERNAME") == "" | Sys.getenv("Z_PASSWORD") == "") {
     zoltr::zoltar_authenticate(zoltar_connection, "zoltar_demo","Dq65&aP0nIlG")
   } else {
@@ -10,7 +21,14 @@ setup_zoltar_connection <- function() {
 }
 
 
-# helper function to get the zoltar project URL
+#' Get Zoltar project URL
+#' 
+#' @param hub character vector, where the first element indicates the hub
+#' from which to load forecasts. Possible options are "US" and "ECDC"
+#' @param zoltar_connection an authenticated Zoltar connecton
+#' 
+#' @return project url
+#' @export
 get_zoltar_project_url <- function(hub = c("US", "ECDC"), 
                                    zoltar_connection = NULL) {
   # if no zoltar connection was provided, set up a new one
@@ -31,9 +49,15 @@ get_zoltar_project_url <- function(hub = c("US", "ECDC"),
   return(project_url)
 }
 
-
-# add stored information about location name and population by joining the 
-# data with the correct stored location data
+#' Add stored information about location name and population by joining the 
+#' data with the correct stored location data
+#'
+#' @param data data frame to append location data
+#' @param hub character vector, where the first element indicates the hub
+#' from which to load forecasts. Possible options are "US" and "ECDC"
+#' 
+#' @return original data with corresponding location information
+#' @export
 join_with_hub_locations <- function(data, 
                                     hub = c("US", "ECDC")) {
   if (hub[1] == "US") {